# Parametric Episode iterator